fundanl/requirements.txt
amqp
celery
psycopg2-binary
schedule
requests_html